FR_1457/churn_predictions-master/churn_predictions-master/requirements.txt
matplotlib==3.10.3
pandas==2.2.3
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
xgboost==3.0.2